clean to base
