+++ src/inter/ascii_scheme.go
@@ -1,0 +1,133 @@
+package inter
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/Fantom-foundation/go-lachesis/src/hash"
+)
+
+// ASCIIschemeToDAG parses events from ASCII-scheme for test purpose.
+// Use joiners ║ ╬ ╠ ╣ ╫ ╚ ╝ ╩ and optional fillers ─ ═ to draw ASCII-scheme.
+// Result:
+//   - nodes  is an array of node addresses;
+//   - events maps node address to array of its events;
+//   - names  maps human readable name to the event;
+func ASCIIschemeToDAG(scheme string) (
+	nodes []hash.Peer, events map[hash.Peer][]*Event, names map[string]*Event) {
+	// init results
+	events = make(map[hash.Peer][]*Event)
+	names = make(map[string]*Event)
+	// read lines
+	for _, line := range strings.Split(strings.TrimSpace(scheme), "\n") {
+		var (
+			nNames    []string // event-N --> name
+			nCreators []int    // event-N --> creator
+			nLinks    [][]int  // event-N --> parents+1 (negative if link to pre-last event)
+		)
+		// parse line
+		current := 1
+		for _, symbol := range strings.Split(strings.TrimSpace(line), " ") {
+			symbol = strings.TrimSpace(symbol)
+			switch symbol {
+			case "─", "═", "": // skip filler
+				current--
+			case "╠", "║╠", "╠╫": // start new link array with current
+				nLinks = append(nLinks, []int{current})
+			case "║╚", "╚": // start new link array with prev
+				nLinks = append(nLinks, []int{-1 * current})
+			case "╣", "╣║", "╫╣", "╬": // append current to last link array
+				last := len(nLinks) - 1
+				nLinks[last] = append(nLinks[last], current)
+			case "╝║", "╝", "╩╫", "╫╩": // append prev to last link array
+				last := len(nLinks) - 1
+				nLinks[last] = append(nLinks[last], -1*current)
+			case "╫", "║", "║║": // don't mutate link array
+				break
+			default: // it is a event name
+				if _, ok := names[symbol]; ok {
+					panic(fmt.Errorf("event '%s' already exists", symbol))
+				}
+				nCreators = append(nCreators, current-1)
+				nNames = append(nNames, symbol)
+				if len(nLinks) < len(nNames) {
+					nLinks = append(nLinks, []int(nil))
+				}
+			}
+			current++
+		}
+		// make nodes if not enough
+		for i := len(nodes); i < (current - 1); i++ {
+			addr := hash.FakePeer()
+			nodes = append(nodes, addr)
+			events[addr] = nil
+		}
+		// create events
+		for i, name := range nNames {
+			// find creator
+			creator := nodes[nCreators[i]]
+			// find creator's parent
+			var (
+				index      uint64
+				parents    = hash.Events{}
+				maxLamport Timestamp
+			)
+			if last := len(events[creator]) - 1; last >= 0 {
+				parent := events[creator][last]
+				index = parent.Index + 1
+				parents.Add(parent.Hash())
+				maxLamport = parent.LamportTime
+			} else {
+				index = 1
+				parents.Add(hash.ZeroEvent)
+				maxLamport = 0
+			}
+			// find other parents
+			for _, p := range nLinks[i] {
+				prev := 0
+				if p < 0 {
+					p *= -1
+					prev = -1
+				}
+				p = p - 1
+				other := nodes[p]
+				last := len(events[other]) - 1 + prev
+				parent := events[other][last]
+				parents.Add(parent.Hash())
+				if maxLamport < parent.LamportTime {
+					maxLamport = parent.LamportTime
+				}
+			}
+			// save event
+			e := &Event{
+				Index:       index,
+				Creator:     creator,
+				Parents:     parents,
+				LamportTime: maxLamport + 1,
+			}
+			events[creator] = append(events[creator], e)
+			names[name] = e
+			hash.EventNameDict[e.Hash()] = name
+		}
+	}
+
+	// human readable names for nodes in log
+	for node, ee := range events {
+		if len(ee) < 1 {
+			continue
+		}
+		name := ee[0].Hash().String()
+		hash.NodeNameDict[node] = "node" + strings.ToUpper(name[0:1])
+	}
+
+	return
+}
+
+// DAGtoASCIIcheme builds ASCII-scheme of events for debug purpose.
+func DAGtoASCIIcheme(events Events) string {
+	events = events.ByParents()
+
+	// TODO: implement it
+
+	return "is not implemented yet"
+}
