+++ src/inter/ascii_scheme_test.go
@@ -1,0 +1,89 @@
+package inter
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/Fantom-foundation/go-lachesis/src/hash"
+)
+
+func TestASCIIschemeToDAG(t *testing.T) {
+	assert := assert.New(t)
+
+	nodes, events, names := ASCIIschemeToDAG(`
+a00 b00   c00 d00
+║   ║     ║   ║
+a01 ║     ║   ║
+║   ╠  ─  c01 ║
+a02 ╣     ║   ║
+║   ║     ║   ║
+╠ ─ ╫ ─ ─ c02 ║
+║   b01  ╝║   ║
+║   ╠ ─ ─ ╫ ─ d01
+║   ║     ║   ║
+║   ║     ║   ║
+╠ ═ b02 ═ ╬   ╣
+║   ║     ║  ║║
+a03 ╣     ╠ ─ d02
+║║  ║     ║  ║║
+║║  ║     ║  ║╠ ─ e00
+║║  ║     ║   ║   ║
+a04 ╫ ─ ─ ╬  ╝║   ║
+║║  ║     ║   ║   ║
+║╚  ╫╩  ─ c03 ╣   ║
+║   ║     ║   ║   ║
+`)
+	expected := map[string][]string{
+		"a00": {""},
+		"a01": {"a00"},
+		"a02": {"a01", "b00"},
+		"a03": {"a02", "b02"},
+		"a04": {"a03", "c02", "d01"},
+		"b00": {""},
+		"b01": {"b00", "c01"},
+		"b02": {"b01", "a02", "c02", "d01"},
+		"c00": {""},
+		"c01": {"c00", "b00"},
+		"c02": {"c01", "a02"},
+		"c03": {"c02", "a03", "b01", "d02"},
+		"d00": {""},
+		"d01": {"d00", "b01"},
+		"d02": {"d01", "c02"},
+		"e00": {"", "d02"},
+	}
+
+	if !assert.Equal(5, len(nodes), "node count") {
+		return
+	}
+	if !assert.Equal(len(expected), len(names), "event count") {
+		return
+	}
+
+	index := make(map[hash.Event]*Event)
+	for _, nodeEvents := range events {
+		for _, e := range nodeEvents {
+			index[e.Hash()] = e
+		}
+	}
+
+	for eName, e := range names {
+		parents := expected[eName]
+		if !assert.Equal(len(parents), len(e.Parents), "at event "+eName) {
+			return
+		}
+		for _, pName := range parents {
+			hash := hash.ZeroEvent
+			if pName != "" {
+				hash = names[pName].Hash()
+			}
+			if !e.Parents.Contains(hash) {
+				t.Fatalf("%s has no parent %s", eName, pName)
+			}
+		}
+	}
+}
+
+func TestDAGtoASCIIcheme(t *testing.T) {
+
+}
